cleanup test_fiat.py
